vq-log: making mypy happy.
Mypy assumes the type of variable to be the first that it sees. If it sees a variable that is assigned a derived class then it assumes derived class type as the variable's type. Hence, it is better to actually assign the uppermost type (or the interface) first to the variable to let mypy know about the variable type being the base-interface rather than the derived class.