std-log: dynamically calculate levels on envLC.
This kind of change needed to be done because `ListLoggerConfigurator.configure()` uses `self.level_list` to get the list of supplied levels and this list needs to be dynamic instead of initialising in the ctor, when we need to read env vars.

This also facilitates users to add certain other env vars at any level list position even after the ctor was called.